started a few things
(DID NOT FINISH EVERYTHING I SAID I WOULD.)
Assigned Cart to Customer that are not in Order.
Made a file to hold lines to delete everything for convenience.
The features file should hold everything else. Currently a working procedure to create Order using Cart, and make assign a new Cart for Customer.